Implement hyper optimized bit shift for light data byte to hold two values

+++ src/engine/network/BlockPlaceUpdate.java
@@ -4,14 +4,14 @@ import org.joml.Vector3i;
 
 public class BlockPlaceUpdate {
     public Vector3i pos;
-    public int ID;
+    public byte ID;
     public byte rot;
 
     public BlockPlaceUpdate(){
 
     }
 
-    public BlockPlaceUpdate(Vector3i newReceivedPos, int ID, byte rot){
+    public BlockPlaceUpdate(Vector3i newReceivedPos, byte ID, byte rot){
         this.pos = newReceivedPos;
         this.ID = ID;
         this.rot = rot;
