+++ src/game/blocks/BlockDefinition.java
@@ -888,7 +888,7 @@ public class BlockDefinition {
             public void onPlace(Vector3d pos) {
                 if (getBlock((int)pos.x, (int)pos.y - 1, (int)pos.z) == 0) {
                     digBlock((int) pos.x, (int) pos.y, (int) pos.z);
-                    addFallingEntity(new Vector3d(pos.x + 0.5d, pos.y, pos.z + 0.5d), new Vector3f(0, 0, 0), 20);
+                    addFallingEntity(new Vector3d(pos.x + 0.5d, pos.y, pos.z + 0.5d), new Vector3f(0, 0, 0), (byte)20);
                 }
             }
         };
@@ -942,7 +942,7 @@ public class BlockDefinition {
                     @Override
                     public void onDig(Vector3d pos) {
                         if (getBlock((int)pos.x, (int)pos.y - 1, (int)pos.z) == 22) {
-                            setBlock((int)pos.x, (int)pos.y - 1, (int)pos.z, 0, 0);
+                            setBlock((int)pos.x, (int)pos.y - 1, (int)pos.z, (byte) 0, 0);
                             createItem("door", pos.add(0.5d,0.5d,0.5d), 1);
                         }
                     }
@@ -951,8 +951,8 @@ public class BlockDefinition {
                     public void onRightClick(Vector3d pos) {
                         if (getBlock((int)pos.x, (int)pos.y - 1, (int)pos.z) == 22) {
                             byte rot = getBlockRotation((int)pos.x, (int)pos.y, (int)pos.z);
-                            setBlock((int)pos.x, (int)pos.y, (int)pos.z, 23,rot);
-                            setBlock((int)pos.x, (int)pos.y - 1, (int)pos.z, 24,rot);
+                            setBlock((int)pos.x, (int)pos.y, (int)pos.z, (byte) 23,rot);
+                            setBlock((int)pos.x, (int)pos.y - 1, (int)pos.z, (byte) 24,rot);
                             playSound("door_close", new Vector3d(pos.x + 0.5d, pos.y + 0.5d, pos.z + 0.5d));
                         }
                     }
@@ -989,7 +989,7 @@ public class BlockDefinition {
                     @Override
                     public void onDig(Vector3d pos) {
                         if (getBlock((int)pos.x, (int)pos.y + 1, (int)pos.z) == 21) {
-                            setBlock((int)pos.x, (int)pos.y + 1, (int)pos.z, 0, 0);
+                            setBlock((int)pos.x, (int)pos.y + 1, (int)pos.z, (byte) 0, 0);
                             createItem("door", pos.add(0.5d,0.5d,0.5d), 1);
                         }
                     }
@@ -998,8 +998,8 @@ public class BlockDefinition {
                     public void onRightClick(Vector3d pos) {
                         if (getBlock((int)pos.x, (int)pos.y + 1, (int)pos.z) == 21) {
                             byte rot = getBlockRotation((int)pos.x, (int)pos.y, (int)pos.z);
-                            setBlock((int)pos.x, (int)pos.y + 1, (int)pos.z, 23,rot);
-                            setBlock((int)pos.x, (int)pos.y, (int)pos.z, 24,rot);
+                            setBlock((int)pos.x, (int)pos.y + 1, (int)pos.z, (byte) 23,rot);
+                            setBlock((int)pos.x, (int)pos.y, (int)pos.z, (byte) 24,rot);
                             playSound("door_close", new Vector3d(pos.x + 0.5d, pos.y + 0.5d, pos.z + 0.5d));
                         }
                     }
@@ -1036,7 +1036,7 @@ public class BlockDefinition {
                     @Override
                     public void onDig(Vector3d pos) {
                         if (getBlock((int)pos.x, (int)pos.y - 1, (int)pos.z) == 24) {
-                            setBlock((int)pos.x, (int)pos.y - 1, (int)pos.z, 0, 0);
+                            setBlock((int)pos.x, (int)pos.y - 1, (int)pos.z, (byte) 0, 0);
                             createItem("door", pos.add(0.5d,0.5d,0.5d), 1);
                         }
                     }
@@ -1045,8 +1045,8 @@ public class BlockDefinition {
                     public void onRightClick(Vector3d pos) {
                         if (getBlock((int)pos.x, (int)pos.y - 1, (int)pos.z) == 24) {
                             byte rot = getBlockRotation((int)pos.x, (int)pos.y, (int)pos.z);
-                            setBlock((int)pos.x, (int)pos.y, (int)pos.z, 21,rot);
-                            setBlock((int)pos.x, (int)pos.y - 1, (int)pos.z, 22,rot);
+                            setBlock((int)pos.x, (int)pos.y, (int)pos.z, (byte) 21,rot);
+                            setBlock((int)pos.x, (int)pos.y - 1, (int)pos.z, (byte) 22,rot);
                             playSound("door_open", new Vector3d(pos.x + 0.5d, pos.y + 0.5d, pos.z + 0.5d));
                         }
                     }
@@ -1083,7 +1083,7 @@ public class BlockDefinition {
                     @Override
                     public void onDig(Vector3d pos) {
                         if (getBlock((int)pos.x, (int)pos.y + 1, (int)pos.z) == 23) {
-                            setBlock((int)pos.x, (int)pos.y + 1, (int)pos.z, 0, 0);
+                            setBlock((int)pos.x, (int)pos.y + 1, (int)pos.z, (byte) 0, 0);
                             createItem("door", pos.add(0.5d,0.5d,0.5d), 1);
                         }
                     }
@@ -1092,8 +1092,8 @@ public class BlockDefinition {
                     public void onRightClick(Vector3d pos) {
                         if (getBlock((int)pos.x, (int)pos.y + 1, (int)pos.z) == 23) {
                             byte rot = getBlockRotation((int)pos.x, (int)pos.y, (int)pos.z);
-                            setBlock((int)pos.x, (int)pos.y + 1, (int)pos.z, 21,rot);
-                            setBlock((int)pos.x, (int)pos.y, (int)pos.z, 22,rot);
+                            setBlock((int)pos.x, (int)pos.y + 1, (int)pos.z, (byte) 21,rot);
+                            setBlock((int)pos.x, (int)pos.y, (int)pos.z, (byte) 22,rot);
                             playSound("door_open", new Vector3d(pos.x + 0.5d, pos.y + 0.5d, pos.z + 0.5d));
                         }
                     }
