+++ src/game/falling/FallingEntityObject.java
@@ -9,9 +9,9 @@ public class FallingEntityObject {
     public Vector3f inertia;
     public Mesh mesh;
     public int key;
-    public int ID;
+    public byte ID;
 
-    public FallingEntityObject(Vector3d pos, Vector3f inertia, Mesh mesh, int key, int ID){
+    public FallingEntityObject(Vector3d pos, Vector3f inertia, Mesh mesh, int key, byte ID){
         this.pos = pos;
         this.inertia = inertia;
         this.mesh = mesh;
